% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{absent}
\alias{absent}
\title{list packages absent from current installation}
\usage{
absent(pset)
}
\arguments{
\item{pset}{PackageSet instance}
}
\description{
list packages absent from current installation
}
