% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{add_dependencies,PackageSet-method}
\alias{add_dependencies,PackageSet-method}
\alias{add_dependencies}
\title{method for adding dependencies to PackageSet as determined by BiocPkgTools}
\usage{
\S4method{add_dependencies}{PackageSet}(pkgset, deps = full_dep_opts(), omit = "R")
}
\arguments{
\item{pkgset}{PackageSet instance}

\item{deps}{character() vector of dependency types (e.g., c("Imports", "Depends")); defaults to \code{full_dep_opts()}}

\item{omit}{character() vector of 'Depends' options to omit -- for example (default) "R"}
}
\description{
method for adding dependencies to PackageSet as determined by BiocPkgTools
}
