% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{bbs_check_built}
\alias{bbs_check_built}
\title{not exported, rely instead on rcmdcheck
do check_built and clean up}
\usage{
bbs_check_built(x)
}
\arguments{
\item{x}{path to package tarball}
}
\description{
not exported, rely instead on rcmdcheck
do check_built and clean up
}
