% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc_to_dataframes.R
\name{bcclist_to_dataframes}
\alias{bcclist_to_dataframes}
\title{operate on BiocCheck return values in a list}
\usage{
bcclist_to_dataframes(bcclist, clean_try_errors = TRUE)
}
\arguments{
\item{bcclist}{list() of outputs of BiocCheck::BiocCheck}

\item{clean_try_errors}{logical(1) if TRUE, scan for try-error and omit those elements}
}
\value{
list of data.frame with basic (pkg, version), notes, warnings, errors from BiocCheck
}
\description{
operate on BiocCheck return values in a list
}
\note{
If x$metadata$Package is not character, element is dropped.
}
