% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bcc.R
\name{bco2df}
\alias{bco2df}
\title{convert BiocCheck output (just error and warning) to list of data.frames}
\usage{
bco2df(x)
}
\arguments{
\item{x}{BiocCheck result}
}
\value{
list of data frames errors, warnings
}
\description{
convert BiocCheck output (just error and warning) to list of data.frames
}
