% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioc_3.10_software_2020.03.17}
\alias{bioc_3.10_software_2020.03.17}
\title{a vector of package names needed for a fairly complete BBS run, 17 march 2020}
\format{
character vector
}
\usage{
bioc_3.10_software_2020.03.17
}
\description{
a vector of package names needed for a fairly complete BBS run, 17 march 2020
}
\examples{
head(bioc_3.10_software_2020.03.17)
}
\keyword{datasets}
