% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{bioc_coreset}
\alias{bioc_coreset}
\title{define a small collection of packages of interest}
\usage{
bioc_coreset(small = TRUE)
}
\arguments{
\item{small}{logical(1) return only 4 packages}
}
\value{
a character vector
}
\description{
define a small collection of packages of interest
}
\examples{
bioc_coreset()
}
