% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build1.R
\name{build1_with_buildsink}
\alias{build1_with_buildsink}
\title{run build1 and capture the R CMD build log}
\usage{
build1_with_buildsink(srcpath, dest = ".", ...)
}
\arguments{
\item{srcpath}{character(1) path to source folder for a package}

\item{dest}{character(1) destination folder}

\item{\dots}{pass to pkgbuild::build}
}
\description{
run build1 and capture the R CMD build log
}
