% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sqlite.R
\name{build_sqlite_db}
\alias{build_sqlite_db}
\title{use outputs of get_checks and get_bcc to build a database of all events of error or warning}
\usage{
build_sqlite_db(
  dbname = "chks.sqlite",
  rcmdcheck_obj_dir,
  bccheck_obj_dir,
  exists_ok = FALSE
)
}
\arguments{
\item{dbname}{character(1) a sqlite database name}

\item{rcmdcheck_obj_dir}{character(1) path to a collection of rcmdcheck outputs}

\item{bccheck_obj_dir}{character(1) path to a collection of bco2df outputs, produced after running
BiocCheck on sources}

\item{exists_ok}{logical(1) passed to \code{status_db_init}}
}
\description{
use outputs of get_checks and get_bcc to build a database of all events of error or warning
}
