% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_PackageSet.R
\name{check_PackageSet}
\alias{check_PackageSet}
\title{longrunning R CMD check and BiocCheck of every package in a packageset}
\usage{
check_PackageSet(
  pset,
  srcdir,
  update_sources = FALSE,
  prefix = tempdir("checkPS"),
  BPPARAM,
  BPOPTIONS = list(),
  shuffle = FALSE,
  sqlite_target = paste0(prefix, "/test.sqlite")
)
}
\arguments{
\item{pset}{BiocBuildTools PackageSet instance}

\item{srcdir}{character(1) path to folder that can receive all sources using BiocBuildTools::populate_local_gits}

\item{update_sources}{logical(1) if TRUE, use update_gits to pull new changes in each package in srcdir}

\item{prefix}{character(1) path to folder within which check artifacts are produced,
will be created if nonexistent; defaults to \code{tempdir("checkPS")}}

\item{BPPARAM}{instance of BiocParallelParam}

\item{BPOPTIONS}{a list, see BiocParallel::bpoptions, defaults to empty list}

\item{shuffle}{logical(1) if TRUE, the list of packages is randomly permuted to avoid always
restarting on the same set, defaults to FALSE}

\item{sqlite_target}{path to SQLite database to be built using BiocBuildTools::build_sqlite_db}
}
\value{
a list with elements sqlite_target, chk, and dbatt which are strings or try-error or
outputs of BiocBuildTools::get_checks2() and BiocBuildTools::build_sqlite_db() if these
succeed.  In good circumstances this function is called for the side effect of creating
the SQLite database
}
\description{
longrunning R CMD check and BiocCheck of every package in a packageset
}
\note{
The function will test Sys.getenv("CI") to avoid a tcltk query from knitr.
}
\examples{
pset = PackageSet(c("parody", "eds"))
srcdir = tempdir("demo_srcs")
spar = BiocParallel::SnowParam(2)
bplog(spar) = TRUE
bpstopOnError(spar) = FALSE
dir.create(ldir <- tempdir("chkps_logs"))
system(paste("chmod 777", ldir))
bplogdir(spar) = tempdir("chkps_logs")
cptry = check_PackageSet(pset, srcdir=srcdir, BPPARAM=spar,
BPOPTIONS=list(exports=c("chkdest", "bdest", "bobdest")))
cptry
}
