% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_status_dbs.R
\name{combine_status_dbs}
\alias{combine_status_dbs}
\title{add records from 'incoming' database to 'target' database}
\usage{
combine_status_dbs(target, incoming)
}
\arguments{
\item{target}{character(1) path to SQLite database to receive new records}

\item{incoming}{character(1) path to SQLite database with new records}
}
\description{
add records from 'incoming' database to 'target' database
}
