% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_tab_demo}
\alias{cov_tab_demo}
\title{a list of covr::package_coverage results, thinned for demonstration purposes, produced dec 2019}
\format{
list
}
\usage{
cov_tab_demo
}
\description{
a list of covr::package_coverage results, thinned for demonstration purposes, produced dec 2019
}
\examples{
cov_tab_demo[1:2]
}
\keyword{datasets}
