% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_rcmdcheck_out}
\alias{demo_rcmdcheck_out}
\title{a list of rcmdcheck::rcmdcheck results for demonstration purposes, produced nov 2022}
\format{
list
}
\usage{
demo_rcmdcheck_out
}
\description{
a list of rcmdcheck::rcmdcheck results for demonstration purposes, produced nov 2022
}
\examples{
names(demo_rcmdcheck_out)
}
\keyword{datasets}
