% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_rcmdchk_out}
\alias{demo_rcmdchk_out}
\title{a list of rcmdcheck::rcmdcheck results for demonstration purposes, produced dec 2019}
\format{
list
}
\usage{
demo_rcmdchk_out
}
\description{
a list of rcmdcheck::rcmdcheck results for demonstration purposes, produced dec 2019
}
\examples{
names(demo_rcmdchk_out)
}
\keyword{datasets}
