% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_events.R
\name{display_events}
\alias{display_events}
\title{simple app to present data on events gathered into a SQLite database}
\usage{
display_events(dbpath)
}
\arguments{
\item{dbpath}{character(1) SQLite database produced by BiocBuildTools::build_sqlite_db}
}
\description{
simple app to present data on events gathered into a SQLite database
}
