% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulfill_deps.R
\name{fulfill_deps}
\alias{fulfill_deps}
\title{use BiocManager::install to get any unavailable dependencies as defined by BiocPkgTools}
\usage{
fulfill_deps(pkg, depdf = NULL, exclude = "KEGG.db", ...)
}
\arguments{
\item{pkg}{character(1) name of package}

\item{depdf}{a data.frame like the output of \code{BiocPkgTools::buildPkgDependencyDataFrame()}}

\item{exclude}{character() of uninstalled packages to be ignored -- BiocManager::install is slow
to deal with an unavailable package}

\item{\dots}{passed to BiocManager::install}
}
\description{
use BiocManager::install to get any unavailable dependencies as defined by BiocPkgTools
}
\note{
Function is used only for side-effect of installation.  \code{exclude} defaults to
'KEGG.db' because in Bioc 3.13 that is not available.  It is understood that code
that performs installation is poor style for scientific programming; this package
concerns software ecosystem management and must be able to programmatically
define the ecosystem for evaluation.
}
