% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{full_dep_opts}
\alias{full_dep_opts}
\title{a vector listing the key dependencies with correct orthography}
\usage{
full_dep_opts()
}
\value{
character vector
}
\description{
a vector listing the key dependencies with correct orthography
}
\examples{
full_dep_opts()
}
