% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checks.R
\name{get_checks}
\alias{get_checks}
\title{run rcmdcheck on all sources associated with a PackageSet and drop the RDS outputs in a folder}
\usage{
get_checks(
  pkgset,
  sources.folder,
  checks.destination,
  BPPARAM = bpparam(),
  BPOPTIONS = bpoptions()
)
}
\arguments{
\item{pkgset}{instance of PackageSet}

\item{sources.folder}{path to git repos}

\item{checks.destination}{path to folder holding RDS of rcmdcheck output}

\item{BPPARAM}{defaults to bpparam()}

\item{BPOPTIONS}{defaults to bpoptions()}
}
\description{
run rcmdcheck on all sources associated with a PackageSet and drop the RDS outputs in a folder
}
\examples{
ps = PackageSet("parody")
tf = tempfile()
td = dir.create(tf)
populate_local_gits(ps, tf)
dir.create(rcdest <- tempfile("rcd"))
m = get_checks(ps, tf, rcdest)
dir(rcdest, full=TRUE)
}
