% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{get_checks2}
\alias{get_checks2}
\title{run rcmdcheck and BiocCheck on all sources associated with a PackageSet and drop the RDS outputs in separate folders}
\usage{
get_checks2(
  pkgset,
  sources.folder,
  checks.destination,
  bcchecks.destination,
  bcobj.destination,
  shuffle.pks = FALSE,
  BPPARAM = bpparam(),
  BPOPTIONS = bpoptions()
)
}
\arguments{
\item{pkgset}{instance of PackageSet}

\item{sources.folder}{path to git repos}

\item{checks.destination}{path to folder holding RDS of rcmdcheck output}

\item{bcchecks.destination}{character(1) path to a writeable folder where BiocCheck logs are written}

\item{bcobj.destination}{character(1) path to a folder where RDS files with data frames are written}

\item{shuffle.pks}{logical(1) if TRUE the packages are randomized before checking}

\item{BPPARAM}{defaults to bpparam()}

\item{BPOPTIONS}{defaults to bpoptions()}
}
\description{
run rcmdcheck and BiocCheck on all sources associated with a PackageSet and drop the RDS outputs in separate folders
}
