% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{get_pnet}
\alias{get_pnet}
\title{use pkgnet to analyze a source folder (including covr statistics)}
\usage{
get_pnet(sourcedir, serialize = TRUE, overwrite = TRUE)
}
\arguments{
\item{sourcedir}{character(1) should contain an R package}

\item{serialize}{logical(1) defaults to TRUE, See note.}

\item{overwrite}{logical(1) defaults to TRUE; if FALSE will fail if \verb{[basename(sourcedir)]_pnet.rds} exists}
}
\value{
See Note.
}
\description{
use pkgnet to analyze a source folder (including covr statistics)
}
\note{
By default, serializes an S3 instance of \code{BBT_pkgnet}, a list with components \code{dep} for dependency report,
and \code{fun} for function interrelationships, saved as RDS to \verb{[basename(sourcedir)]_pnet.rds}.  If
\code{serialize} is FALSE, simply returns the object.  This is intended to be used
on a large number of packages and thus a checkpointing approach is the default.
}
