% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpk.R
\name{getpk}
\alias{getpk}
\title{use HTTPS to clone a package from Bioconductor git}
\usage{
getpk(x)
}
\arguments{
\item{x}{character(1) package name expected to be a Bioconductor package in git}
}
\value{
result of system()
}
\description{
use HTTPS to clone a package from Bioconductor git
}
\examples{
td = tempdir()
wd = getwd()
setwd(td)
lk = getpk("parody")
if (FALSE) lk2 = try(getpk("parody_z")) # should fail
setwd(wd)
}
