% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{installed_r}
\alias{installed_r}
\title{a vector of package names needed for a fairly complete BBS run, jul 7 2019}
\format{
character vector
}
\usage{
installed_r
}
\description{
a vector of package names needed for a fairly complete BBS run, jul 7 2019
}
\examples{
data(installed_r)
head(installed_r)
}
\keyword{datasets}
