% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checks.R
\name{last_commit_date}
\alias{last_commit_date}
\title{use gert to get last commit date for a repo}
\usage{
last_commit_date(repo)
}
\arguments{
\item{repo}{character(1) path to a checkout from git}
}
\description{
use gert to get last commit date for a repo
}
