% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{local_gits_with_uninstalled_dependencies}
\alias{local_gits_with_uninstalled_dependencies}
\title{not clear that this is needed
provide a list of packages for which dependencies are not installed}
\usage{
local_gits_with_uninstalled_dependencies(
  gitspath,
  dependencies = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{gitspath}{character(1) path to cloned repos}

\item{dependencies}{character vector of dependency types, probably don't need "Suggests"}
}
\description{
not clear that this is needed
provide a list of packages for which dependencies are not installed
}
