% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet.R
\name{make_pnet_object}
\alias{make_pnet_object}
\title{make a serialized collection of pkgnet reports}
\usage{
make_pnet_object(src_folders, target = NULL)
}
\arguments{
\item{src_folders}{character() collection of source package folders}

\item{target}{optional character(1) defaults to NULL.  If non-null, a saveRDS will be performed.}
}
\value{
list, each element has components 'dep' and 'fun' corresponding to DependencyReporter
and FunctionReporter for each package.  The list is classed as BBT_pkgnet for printing.
}
\description{
make a serialized collection of pkgnet reports
}
\examples{
if (interactive()) {
 folds = dir(system.file("demo_srcs", package="BiocBuildTools"), full.names=TRUE)
 tf = tempfile()
 y = make_pnet_object(folds, target=tf)
 x = readRDS(tf)
 x$parody$fun$graph_viz
 y
}
}
