% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet.R
\name{print.BBT_pkgnet}
\alias{print.BBT_pkgnet}
\title{compact view of a pkgnet list}
\usage{
\method{print}{BBT_pkgnet}(x, ...)
}
\arguments{
\item{x}{BBT_pkgnet instance}

\item{\dots}{not used}
}
\description{
compact view of a pkgnet list
}
