% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_artifacts.R
\name{process_artifacts}
\alias{process_artifacts}
\title{process outputs of tbchecks and get_pnet to run browse_checks}
\usage{
process_artifacts(
  dbpath = tempfile(),
  run_app = TRUE,
  startup = TRUE,
  serialize_pnet = TRUE,
  pnet_obj_name = "pnet.rds"
)
}
\arguments{
\item{dbpath}{character(1) a SQLite database will be produced here}

\item{run_app}{logical(1) defaults to TRUE; runs \code{browse_checks} on available data}

\item{startup}{logical(1) defaults to TRUE; passed to \code{status_db_init}}

\item{serialize_pnet}{logical(1) if TRUE emits list of pkgnet outputs as RDS}

\item{pnet_obj_name}{character(1) defaults to "pnet.rds"}
}
\description{
process outputs of tbchecks and get_pnet to run browse_checks
}
\note{
Starts a SQLite connection.  The connection is closed in the call to dbConnect, this
should be conditioned with an on.exit here in case the connection was not closed.
If run_app is FALSE, the connection is closed on exit.
This function will create a SQLite database and list of pkgnet objects for use
with browse_checks.
}
