% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bcc.R
\name{process_bcc}
\alias{process_bcc}
\title{process BiocCheck logs into a single data.frame emphasizing non-OK lines}
\usage{
process_bcc(folder)
}
\arguments{
\item{folder}{character(1) path to a collection of BiocCheck outputs (packagename/00BiocCheck.log)}
}
\description{
process BiocCheck logs into a single data.frame emphasizing non-OK lines
}
