% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sqlite.R
\name{rcclist_to_dataframes}
\alias{rcclist_to_dataframes}
\title{produce list of data.frames for storage of R CMD check results in SQLite}
\usage{
rcclist_to_dataframes(rcclist)
}
\arguments{
\item{rcclist}{list of results of rcmdcheck::rcmdcheck}
}
\description{
produce list of data.frames for storage of R CMD check results in SQLite
}
\examples{
data(demo_rcmdchk_out)
dfs = rcclist_to_dataframes(demo_rcmdchk_out)
names(dfs)
names(dfs[[1]])
}
