% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_rcmdcheck.R
\name{safe_rcmdcheck}
\alias{safe_rcmdcheck}
\title{deal with errors at \emph{build} stage when attempting to check}
\usage{
safe_rcmdcheck(x, error = "never", ...)
}
\arguments{
\item{x}{character(1) path to source package folder for a package (not a tarball)}

\item{error}{character() passed to rcmdcheck::rcmdcheck}

\item{\dots}{passed to rcmdcheck::rcmdcheck, should not include 'error'}
}
\value{
a list, see note.
}
\description{
deal with errors at \emph{build} stage when attempting to check
}
\note{
If rcmdcheck succeeds without returning \code{try-error}, this function returns
the result of rcmdcheck.  Otherwise it build a list with elements 'package',
'errors', 'warnings', 'notes', 'install_out', 'description' and 'version',
in which the 'errors' component includes data from \code{capture.output} on
rcmdcheck attempt.
}
