% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soft_021520}
\alias{soft_021520}
\title{a vector of package names needed for a fairly complete BBS run, 15 feb 2020}
\format{
character vector
}
\usage{
soft_021520
}
\description{
a vector of package names needed for a fairly complete BBS run, 15 feb 2020
}
\examples{
head(soft_021520)
}
\keyword{datasets}
