% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_db.R
\name{status_db_init}
\alias{status_db_init}
\title{initialize a status database}
\usage{
status_db_init(sqlite_filename, named_df_list, exists_ok = FALSE)
}
\arguments{
\item{sqlite_filename}{character(1) typically new path}

\item{named_df_list}{a list of data.frame instances}

\item{exists_ok}{logical(1) if TRUE will work with existing db}
}
\description{
initialize a status database
}
