% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{tbcheck}
\alias{tbcheck}
\title{check an R package tarball}
\usage{
tbcheck(
  x,
  error = "never",
  args = "--no-build-vignettes",
  convert = basename,
  ...
)
}
\arguments{
\item{x}{character(1) path to tarball}

\item{error}{passed to \code{rcmdcheck::rcmdcheck}}

\item{args}{passed to \code{rcmdcheck::rcmdcheck}}

\item{convert}{function that converts path to a package name}

\item{\dots}{passed to \code{rcmdcheck::rcmdcheck}}
}
\description{
check an R package tarball
}
\note{
writes objects \verb{[pname]_chk.rds} and \verb{[pname]_bcchk.rds} to disk
for later use
td = tempdir()
cur = getwd()
setwd(td)
pdy = system.file("demo_srcs/parody", package="BiocBuildTools")
pkgbuild::build(pdy)
tbcheck(dir(patt="parody_"))
file.exists("parody_chk.rds")
setwd(cur)
}
