% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gits.R
\name{update_gits}
\alias{update_gits}
\title{attempt to identify packages that have had pushes since last collected}
\usage{
update_gits(folder = "~/BBS_space/NOV5_FULL/fullgits_nov10")
}
\arguments{
\item{folder}{character(1) folder in which checked out packages reside}
}
\value{
a vector of package names for which changes were observed.
}
\description{
attempt to identify packages that have had pushes since last collected
}
\note{
We use setwd(folder) because it seems we need to use system('git pull').
The principle is that \code{gert::git_pull} returns an error when there has been a change. (?)
When this happens, we enter the folder and use system('git pull') to get the changes.
}
