% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status_db.R
\name{update_status_db}
\alias{update_status_db}
\title{add a dataframe, optionally purging related obsolete records, to a SQLite database}
\usage{
update_status_db(
  sqlite_filename,
  tablename,
  dropfield = NULL,
  dropvalue,
  newdf
)
}
\arguments{
\item{sqlite_filename}{character(1) filename of SQLite database; database must exist}

\item{tablename}{character(1) name of table to operate on}

\item{dropfield}{character(1) or NULL, the field used to select records to delete}

\item{dropvalue}{character(1) only used if \code{dropfield} is not NULL, value used to
select records for deletion}

\item{newdf}{data.frame conforming to the table in \code{tablename} if such exists}
}
\description{
add a dataframe, optionally purging related obsolete records, to a SQLite database
}
\note{
Connects to db with flag SQLITE_RW and disconnects on exit.
}
