% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bcc_process.R
\name{bcc_to_dataframes}
\alias{bcc_to_dataframes}
\title{now that BiocCheck in 3.15 uses an envRefClass, we must operate on BiocCheck at once}
\usage{
bcc_to_dataframes(folder)
}
\arguments{
\item{folder}{a path to checkout of a package}
}
\description{
now that BiocCheck in 3.15 uses an envRefClass, we must operate on BiocCheck at once
}
\examples{
pks = dir(system.file("demo_srcs", package="BiocBuildTools"), full.names=TRUE)
dfs = lapply(pks, bcc_to_dataframes)
head(dfs[[1]][[1]])
}
