% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sqlite.R
\name{bcclist_to_dataframes}
\alias{bcclist_to_dataframes}
\title{helper function for extracting BiocCheck outputs}
\usage{
bcclist_to_dataframes(bcclist)
}
\arguments{
\item{bcclist}{list of outputs of BiocCheck::BiocCheck}
}
\description{
helper function for extracting BiocCheck outputs
}
\examples{
data(demo_bcchk_out)
bl = bcclist_to_dataframes(demo_bcchk_out)
sapply(bl, dim)
}
