% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcc.R
\name{.bco2df}
\alias{.bco2df}
\title{helper for bco2df}
\usage{
.bco2df(x)
}
\arguments{
\item{x}{the warning or error component of BiocCheck}
}
\value{
data.frame with checkType
}
\description{
helper for bco2df
}
