% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcc.R
\name{get_bcc}
\alias{get_bcc}
\title{use bco2df to acquire limited data frames from BiocCheck applied to packages}
\usage{
get_bcc(
  sources.folder,
  bcchecks.destination,
  bcobj.destination,
  BPPARAM = bpparam(),
  BPOPTIONS = bpoptions()
)
}
\arguments{
\item{sources.folder}{character(1) path to github checkouts of Bioc packages}

\item{bcchecks.destination}{character(1) path to a writeable folder where BiocCheck logs are written}

\item{bcobj.destination}{character(1) path to a folder where RDS files with data frames are written}

\item{BPPARAM}{defaults to bpparam()}

\item{BPOPTIONS}{defaults to bpoptions()}
}
\description{
use bco2df to acquire limited data frames from BiocCheck applied to packages
}
\examples{
ps = PackageSet("parody")
tf = tempfile()
td = dir.create(tf)
populate_local_gits(ps, tf)
dir.create(bcdest <- tempfile("bcd"))
dir.create(bcodest <- tempfile("bco"))
get_bcc(tf, bcdest, bcodest)
readRDS(dir(bcodest, full=TRUE))
\dontrun{
set.seed(1234) # we shuffle packages to avoid restarting on a bad one
bcchecks.destination = dir.create(tempfile("bccheck"))
bcobj.destination = dir.create(tempfile("bcobj"))
spar = BiocParallel::SnowParam(3, threshold="DEBUG")
BiocParallel::bplog(spar) = TRUE
# make 777 BiocParallel::bplogdir(spar) = "~/BBS_space/BPLOGS_bc"
BiocParallel::bptimeout(spar) = 600
BiocParallel::register(spar)
# get_bcc(  ...
}
}
