% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpk.R
\name{getpk}
\alias{getpk}
\title{use HTTPS to clone a package from Bioconductor git}
\usage{
getpk(x, branch = "RELEASE_3_15", nuke_inst_doc = TRUE)
}
\arguments{
\item{x}{character(1) package name expected to be a Bioconductor package in git}

\item{branch}{character(1) "RELEASE_3_15" is default}

\item{nuke_inst_doc}{logical(1) defaults to TRUE ... to avoid pkgbuild::build query, remove inst/doc}
}
\value{
result of system()
}
\description{
use HTTPS to clone a package from Bioconductor git
}
\examples{
td = tempdir()
wd = getwd()
setwd(td)
lk = getpk("parody")
if (FALSE) lk2 = try(getpk("parody_z")) # should fail
setwd(wd)
}
