% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{PackageSet}
\alias{PackageSet}
\title{constructor for PackageSet instances}
\usage{
PackageSet(cvec, biocversion = "3.11")
}
\arguments{
\item{cvec}{character() vector}

\item{biocversion}{character(1) defaulting to "3.11"}
}
\description{
constructor for PackageSet instances
}
\note{
Will issue message if some element of cvec is not
found in BiocPkgTools::biocPkgList() result
}
